requirements.txt
fastapi[all]
mysql-connector-python
uvicorn
requests